Filter a gtfs feed so that it only contains trips running on a given date

Only stop_times, stops, routes, services (in calendar and calendar_dates), shapes, frequencies and transfers belonging to one of those trips are kept.

filter_feed_by_date(
  gtfs_obj,
  extract_date,
  min_departure_time = "00:00:00",
  max_arrival_time = "48:00:00"
)

Arguments

gtfs_obj
a gtfs feed

extract_date
date to extract trips from in YYYY-MM-DD format

min_departure_time
The earliest departure time. Can be given as "HH:MM:SS", hms object or numeric value in seconds.

max_arrival_time
The latest arrival time. Can be given as "HH:MM:SS", hms object or numeric value in seconds

Value

tidygtfs object with filtered tables

See also