Validate GTFS file

Validates the GTFS object against GTFS specifications and raises warnings if required files/fields are not found.

validate_gtfs(gtfs, files = NULL, quiet = TRUE, warnings = TRUE)

Arguments

gtfs
A GTFS object as created by read_gtfs.

files
A character vector containing the text files to be validated against the GTFS specification (without the .txt extension). If NULL (the default) the provided GTFS is validated against all possible GTFS text files.

quiet
Whether to hide log messages (defaults to TRUE).

warnings
Whether to display warning messages (defaults to TRUE).

Value

A GTFS object with a validation_result attribute. This attribute is a data.table containing the validation summary of all possible fields from the specified files.

Details

GTFS object's files and fields are validated against the GTFS specifications as documented in Google's Static GTFS Reference:

GTFS feeds are considered valid if they include all required files and fields. If a required file/field is missing the function (optionally) raises a warning.

Optional files/fields are listed in the reference above but are not required, thus no warning is raised if they are missing.

Extra files/fields are those who are not listed in the reference above (either because they refer to a specific GTFS extension or due to any other reason).

Note that some files (calendar.txt, calendar_dates.txt and feed_info.txt) are conditionally required. This means that:

calendar.txt is initially set as a required file. If it's not present, however, it becomes optional and calendar_dates.txt (originally set as optional) becomes required.

feed_info.txt is initially set as an optional file. If translations.txt is present, however, it becomes required.

Examples

local_gtfs_path <- system.file("extdata", "google_transit_nyc_subway.zip", package = "tidytransit")
gtfs <- read_gtfs(local_gtfs_path)
attr(gtfs, "validation_result")

#>                file file_spec file_provided_status                  field
#> 1            agency       req                 TRUE              agency_id
#> 2            agency       req                 TRUE            agency_name
#> 3            agency       req                 TRUE             agency_url
#> 4            agency       req                 TRUE        agency_timezone
#> 5            agency       req                 TRUE            agency_lang
#> 6            agency       req                 TRUE           agency_phone
#> 7            agency       req                 TRUE        agency_fare_url
#> 8            agency       req                 TRUE           agency_email
#> 9             stops       req                 TRUE                stop_id
#> 10            stops       req                 TRUE              stop_code
#> 11            stops       req                 TRUE              stop_name
#> 12            stops       req                 TRUE              stop_desc
#> 13            stops       req                 TRUE               stop_lat
#> 14            stops       req                 TRUE               stop_lon
#> 15            stops       req                 TRUE                zone_id
#> 16            stops       req                 TRUE               stop_url
#> 17            stops       req                 TRUE          location_type
#> 18            stops       req                 TRUE         parent_station
#> 19            stops       req                 TRUE          stop_timezone
#> 20            stops       req                 TRUE    wheelchair_boarding
#> 21            stops       req                 TRUE               level_id
#> 22            stops       req                 TRUE          platform_code
#> 23           routes       req                 TRUE               route_id
#> 24           routes       req                 TRUE              agency_id
#> 25           routes       req                 TRUE       route_short_name
#> 26           routes       req                 TRUE        route_long_name
#> 27           routes       req                 TRUE             route_desc
#> 28           routes       req                 TRUE             route_type
#> 29           routes       req                 TRUE              route_url
#> 30           routes       req                 TRUE            route_color
#> 31           routes       req                 TRUE       route_text_color
#> 32           routes       req                 TRUE       route_sort_order
#> 33           routes       req                 TRUE      continuous_pickup
#> 34           routes       req                 TRUE    continuous_drop_off
#> 35            trips       req                 TRUE               route_id
#> 36            trips       req                 TRUE             service_id
#> 37            trips       req                 TRUE                trip_id
#> 38            trips       req                 TRUE          trip_headsign
#> 39            trips       req                 TRUE        trip_short_name
#> 40            trips       req                 TRUE           direction_id
#> 41            trips       req                 TRUE               block_id
#> 42            trips       req                 TRUE               shape_id
#> 43            trips       req                 TRUE  wheelchair_accessible
#> 44            trips       req                 TRUE          bikes_allowed
#> 45       stop_times       req                 TRUE                trip_id
#> 46       stop_times       req                 TRUE           arrival_time
#> 47       stop_times       req                 TRUE         departure_time
#> 48       stop_times       req                 TRUE                stop_id
#> 49       stop_times       req                 TRUE          stop_sequence
#> 50       stop_times       req                 TRUE          stop_headsign
#> 51       stop_times       req                 TRUE            pickup_type
#> 52       stop_times       req                 TRUE          drop_off_type
#> 53       stop_times       req                 TRUE      continuous_pickup
#> 54       stop_times       req                 TRUE    continuous_drop_off
#> 55       stop_times       req                 TRUE    shape_dist_traveled
#> 56       stop_times       req                 TRUE              timepoint
#> 57         calendar       req                 TRUE             service_id
#> 58         calendar       req                 TRUE                 monday
#> 59         calendar       req                 TRUE                tuesday
#> 60         calendar       req                 TRUE              wednesday
#> 61         calendar       req                 TRUE               thursday
#> 62         calendar       req                 TRUE                 friday
#> 63         calendar       req                 TRUE               saturday
#> 64         calendar       req                 TRUE                 sunday
#> 65         calendar       req                 TRUE             start_date
#> 66         calendar       req                 TRUE               end_date
#> 67   calendar_dates       opt                 TRUE             service_id
#> 68   calendar_dates       opt                 TRUE                   date
#> 69   calendar_dates       opt                 TRUE         exception_type
#> 70  fare_attributes       opt                FALSE              agency_id
#> 71  fare_attributes       opt                FALSE                fare_id
#> 72  fare_attributes       opt                FALSE                  price
#> 73  fare_attributes       opt                FALSE          currency_type
#> 74  fare_attributes       opt                FALSE         payment_method
#> 75  fare_attributes       opt                FALSE              transfers
#> 76  fare_attributes       opt                FALSE      transfer_duration
#> 77       fare_rules       opt                FALSE                fare_id
#> 78       fare_rules       opt                FALSE               route_id
#> 79       fare_rules       opt                FALSE              origin_id
#> 80       fare_rules       opt                FALSE         destination_id
#> 81       fare_rules       opt                FALSE            contains_id
#> 82           shapes       opt                 TRUE               shape_id
#> 83           shapes       opt                 TRUE           shape_pt_lat
#> 84           shapes       opt                 TRUE           shape_pt_lon
#> 85           shapes       opt                 TRUE      shape_pt_sequence
#> 86           shapes       opt                 TRUE    shape_dist_traveled
#> 87      frequencies       opt                FALSE                trip_id
#> 88      frequencies       opt                FALSE             start_time
#> 89      frequencies       opt                FALSE               end_time
#> 90      frequencies       opt                FALSE           headway_secs
#> 91      frequencies       opt                FALSE            exact_times
#> 92        transfers       opt                 TRUE           from_stop_id
#> 93        transfers       opt                 TRUE             to_stop_id
#> 94        transfers       opt                 TRUE          transfer_type
#> 95        transfers       opt                 TRUE      min_transfer_time
#> 96         pathways       opt                FALSE             pathway_id
#> 97         pathways       opt                FALSE           from_stop_id
#> 98         pathways       opt                FALSE             to_stop_id
#> 99         pathways       opt                FALSE           pathway_mode
#> 100        pathways       opt                FALSE       is_bidirectional
#> 101        pathways       opt                FALSE                 length
#> 102        pathways       opt                FALSE         traversal_time
#> 103        pathways       opt                FALSE            stair_count
#> 104        pathways       opt                FALSE              max_slope
#> 105        pathways       opt                FALSE              min_width
#> 106        pathways       opt                FALSE          signposted_as
#> 107        pathways       opt                FALSE reversed_signposted_as
#> 108          levels       opt                FALSE               level_id
#> 109          levels       opt                FALSE            level_index
#> 110          levels       opt                FALSE             level_name
#> 111       feed_info       opt                FALSE    feed_publisher_name
#> 112       feed_info       opt                FALSE     feed_publisher_url
#> 113       feed_info       opt                FALSE              feed_lang
#> 114       feed_info       opt                FALSE        feed_start_date
#> 115       feed_info       opt                FALSE          feed_end_date
#> 116       feed_info       opt                FALSE           feed_version
#> 117       feed_info       opt                FALSE     feed_contact_email
#> 118       feed_info       opt                FALSE       feed_contact_url
#> 119    translations       opt                FALSE             table_name
#> 120    translations       opt                FALSE             field_name
#> 121    translations       opt                FALSE               language
#> 122    translations       opt                FALSE            translation
#> 123    translations       opt                FALSE              record_id
#> 124    translations       opt                FALSE          record_sub_id
#> 125    translations       opt                FALSE            field_value
#> 126    attributions       opt                FALSE         attribution_id
#> 127    attributions       opt                FALSE              agency_id
#> 128    attributions       opt                FALSE               route_id
#> 129    attributions       opt                FALSE                trip_id
#> 130    attributions       opt                FALSE      organization_name
#> 131    attributions       opt                FALSE            is_producer
#> 132    attributions       opt                FALSE            is_operator
#> 133    attributions       opt                FALSE           is_authority
#> 134    attributions       opt                FALSE        attribution_url
#> 135    attributions       opt                FALSE      attribution_email
#> 136    attributions       opt                FALSE      attribution_phone
#> 137           trips       ext                 TRUE               route_id
#> 138           trips       ext                 TRUE             service_id
#> 139           trips       ext                 TRUE                trip_id
#> 140           trips       ext                 TRUE          trip_headsign
#> 141           trips       ext                 TRUE           direction_id
#> 142           trips       ext                 TRUE               block_id
#> 143           trips       ext                 TRUE               shape_id
#> 144      stop_times       ext                 TRUE                trip_id
#> 145      stop_times       ext                 TRUE           arrival_time
#> 146      stop_times       ext                 TRUE         departure_time
#> 147      stop_times       ext                 TRUE                stop_id
#> 148      stop_times       ext                 TRUE          stop_sequence
#> 149      stop_times       ext                 TRUE          stop_headsign
#> 150      stop_times       ext                 TRUE            pickup_type
#> 151      stop_times       ext                 TRUE          drop_off_type
#> 152      stop_times       ext                 TRUE    shape_dist_traveled
#> 153          agency       ext                 TRUE              agency_id
#> 154          agency       ext                 TRUE            agency_name
#> 155          agency       ext                 TRUE             agency_url
#> 156          agency       ext                 TRUE        agency_timezone
#> 157          agency       ext                 TRUE            agency_lang
#> 158          agency       ext                 TRUE           agency_phone
#> 159        calendar       ext                 TRUE             service_id
#> 160        calendar       ext                 TRUE                 monday
#> 161        calendar       ext                 TRUE                tuesday
#> 162        calendar       ext                 TRUE              wednesday
#> 163        calendar       ext                 TRUE               thursday
#> 164        calendar       ext                 TRUE                 friday
#> 165        calendar       ext                 TRUE               saturday
#> 166        calendar       ext                 TRUE                 sunday
#> 167        calendar       ext                 TRUE             start_date
#> 168        calendar       ext                 TRUE               end_date
#> 169  calendar_dates       ext                 TRUE             service_id
#> 170  calendar_dates       ext                 TRUE                   date
#> 171  calendar_dates       ext                 TRUE         exception_type
#> 172           stops       ext                 TRUE                stop_id
#> 173           stops       ext                 TRUE              stop_code
#> 174           stops       ext                 TRUE              stop_name
#> 175           stops       ext                 TRUE              stop_desc
#> 176           stops       ext                 TRUE               stop_lat
#> 177           stops       ext                 TRUE               stop_lon
#> 178           stops       ext                 TRUE                zone_id
#> 179           stops       ext                 TRUE               stop_url
#> 180           stops       ext                 TRUE          location_type
#> 181           stops       ext                 TRUE         parent_station
#> 182          routes       ext                 TRUE               route_id
#> 183          routes       ext                 TRUE              agency_id
#> 184          routes       ext                 TRUE       route_short_name
#> 185          routes       ext                 TRUE        route_long_name
#> 186          routes       ext                 TRUE             route_desc
#> 187          routes       ext                 TRUE             route_type
#> 188          routes       ext                 TRUE              route_url
#> 189          routes       ext                 TRUE            route_color
#> 190          routes       ext                 TRUE       route_text_color
#> 191          shapes       ext                 TRUE               shape_id
#> 192          shapes       ext                 TRUE           shape_pt_lat
#> 193          shapes       ext                 TRUE           shape_pt_lon
#> 194          shapes       ext                 TRUE      shape_pt_sequence
#> 195          shapes       ext                 TRUE    shape_dist_traveled
#> 196       transfers       ext                 TRUE           from_stop_id
#> 197       transfers       ext                 TRUE             to_stop_id
#> 198       transfers       ext                 TRUE          transfer_type
#> 199       transfers       ext                 TRUE      min_transfer_time
#>     field_spec field_provided_status validation_status validation_details
#> 1          opt                  TRUE                ok               <NA>
#> 2          req                  TRUE                ok               <NA>
#> 3          req                  TRUE                ok               <NA>
#> 4          req                  TRUE                ok               <NA>
#> 5          opt                  TRUE                ok               <NA>
#> 6          opt                  TRUE                ok               <NA>
#> 7          opt                 FALSE              info  missing_opt_field
#> 8          opt                 FALSE              info  missing_opt_field
#> 9          req                  TRUE                ok               <NA>
#> 10         opt                  TRUE                ok               <NA>
#> 11         req                  TRUE                ok               <NA>
#> 12         opt                  TRUE                ok               <NA>
#> 13         req                  TRUE                ok               <NA>
#> 14         req                  TRUE                ok               <NA>
#> 15         opt                  TRUE                ok               <NA>
#> 16         opt                  TRUE                ok               <NA>
#> 17         opt                  TRUE                ok               <NA>
#> 18         opt                  TRUE                ok               <NA>
#> 19         opt                 FALSE              info  missing_opt_field
#> 20         opt                 FALSE              info  missing_opt_field
#> 21         opt                 FALSE              info  missing_opt_field
#> 22         opt                 FALSE              info  missing_opt_field
#> 23         req                  TRUE                ok               <NA>
#> 24         opt                  TRUE                ok               <NA>
#> 25         req                  TRUE                ok               <NA>
#> 26         req                  TRUE                ok               <NA>
#> 27         opt                  TRUE                ok               <NA>
#> 28         req                  TRUE                ok               <NA>
#> 29         opt                  TRUE                ok               <NA>
#> 30         opt                  TRUE                ok               <NA>
#> 31         opt                  TRUE                ok               <NA>
#> 32         opt                 FALSE              info  missing_opt_field
#> 33         opt                 FALSE              info  missing_opt_field
#> 34         opt                 FALSE              info  missing_opt_field
#> 35         req                  TRUE                ok               <NA>
#> 36         req                  TRUE                ok               <NA>
#> 37         req                  TRUE                ok               <NA>
#> 38         opt                  TRUE                ok               <NA>
#> 39         opt                 FALSE              info  missing_opt_field
#> 40         opt                  TRUE                ok               <NA>
#> 41         opt                  TRUE                ok               <NA>
#> 42         opt                  TRUE                ok               <NA>
#> 43         opt                 FALSE              info  missing_opt_field
#> 44         opt                 FALSE              info  missing_opt_field
#> 45         req                  TRUE                ok               <NA>
#> 46         req                  TRUE                ok               <NA>
#> 47         req                  TRUE                ok               <NA>
#> 48         req                  TRUE                ok               <NA>
#> 49         req                  TRUE                ok               <NA>
#> 50         opt                  TRUE                ok               <NA>
#> 51         opt                  TRUE                ok               <NA>
#> 52         opt                  TRUE                ok               <NA>
#> 53         opt                 FALSE              info  missing_opt_field
#> 54         opt                 FALSE              info  missing_opt_field
#> 55         opt                  TRUE                ok               <NA>
#> 56         opt                 FALSE              info  missing_opt_field
#> 57         req                  TRUE                ok               <NA>
#> 58         req                  TRUE                ok               <NA>
#> 59         req                  TRUE                ok               <NA>
#> 60         req                  TRUE                ok               <NA>
#> 61         req                  TRUE                ok               <NA>
#> 62         req                  TRUE                ok               <NA>
#> 63         req                  TRUE                ok               <NA>
#> 64         req                  TRUE                ok               <NA>
#> 65         req                  TRUE                ok               <NA>
#> 66         req                  TRUE                ok               <NA>
#> 67         req                  TRUE                ok               <NA>
#> 68         req                  TRUE                ok               <NA>
#> 69         req                  TRUE                ok               <NA>
#> 70         opt                 FALSE              info   missing_opt_file
#> 71         req                 FALSE              info   missing_opt_file
#> 72         req                 FALSE              info   missing_opt_file
#> 73         req                 FALSE              info   missing_opt_file
#> 74         req                 FALSE              info   missing_opt_file
#> 75         req                 FALSE              info   missing_opt_file
#> 76         opt                 FALSE              info   missing_opt_file
#> 77         req                 FALSE              info   missing_opt_file
#> 78         opt                 FALSE              info   missing_opt_file
#> 79         opt                 FALSE              info   missing_opt_file
#> 80         opt                 FALSE              info   missing_opt_file
#> 81         opt                 FALSE              info   missing_opt_file
#> 82         req                  TRUE                ok               <NA>
#> 83         req                  TRUE                ok               <NA>
#> 84         req                  TRUE                ok               <NA>
#> 85         req                  TRUE                ok               <NA>
#> 86         opt                  TRUE                ok               <NA>
#> 87         req                 FALSE              info   missing_opt_file
#> 88         req                 FALSE              info   missing_opt_file
#> 89         req                 FALSE              info   missing_opt_file
#> 90         req                 FALSE              info   missing_opt_file
#> 91         opt                 FALSE              info   missing_opt_file
#> 92         req                  TRUE                ok               <NA>
#> 93         req                  TRUE                ok               <NA>
#> 94         req                  TRUE                ok               <NA>
#> 95         opt                  TRUE                ok               <NA>
#> 96         req                 FALSE              info   missing_opt_file
#> 97         req                 FALSE              info   missing_opt_file
#> 98         req                 FALSE              info   missing_opt_file
#> 99         req                 FALSE              info   missing_opt_file
#> 100        req                 FALSE              info   missing_opt_file
#> 101        opt                 FALSE              info   missing_opt_file
#> 102        opt                 FALSE              info   missing_opt_file
#> 103        opt                 FALSE              info   missing_opt_file
#> 104        opt                 FALSE              info   missing_opt_file
#> 105        opt                 FALSE              info   missing_opt_file
#> 106        opt                 FALSE              info   missing_opt_file
#> 107        opt                 FALSE              info   missing_opt_file
#> 108        req                 FALSE              info   missing_opt_file
#> 109        req                 FALSE              info   missing_opt_file
#> 110        opt                 FALSE              info   missing_opt_file
#> 111        req                 FALSE              info   missing_opt_file
#> 112        req                 FALSE              info   missing_opt_file
#> 113        req                 FALSE              info   missing_opt_file
#> 114        opt                 FALSE              info   missing_opt_file
#> 115        opt                 FALSE              info   missing_opt_file
#> 116        opt                 FALSE              info   missing_opt_file
#> 117        opt                 FALSE              info   missing_opt_file
#> 118        opt                 FALSE              info   missing_opt_file
#> 119        req                 FALSE              info   missing_opt_file
#> 120        req                 FALSE              info   missing_opt_file
#> 121        req                 FALSE              info   missing_opt_file
#> 122        req                 FALSE              info   missing_opt_file
#> 123        opt                 FALSE              info   missing_opt_file
#> 124        opt                 FALSE              info   missing_opt_file
#> 125        opt                 FALSE              info   missing_opt_file
#> 126        opt                 FALSE              info   missing_opt_file
#> 127        opt                 FALSE              info   missing_opt_file
#> 128        opt                 FALSE              info   missing_opt_file
#> 129        opt                 FALSE              info   missing_opt_file
#> 130        req                 FALSE              info   missing_opt_file
#> 131        opt                 FALSE              info   missing_opt_file
#> 132        opt                 FALSE              info   missing_opt_file
#> 133        opt                 FALSE              info   missing_opt_file
#> 134        opt                 FALSE              info   missing_opt_file
#> 135        opt                 FALSE              info   missing_opt_file
#> 136        opt                 FALSE              info   missing_opt_file
#> 137        ext                  TRUE              info  undocumented_file
#> 138        ext                  TRUE              info  undocumented_file
#> 139        ext                  TRUE              info  undocumented_file
#> 140        ext                  TRUE              info  undocumented_file
#> 141        ext                  TRUE              info  undocumented_file
#> 142        ext                  TRUE              info  undocumented_file
#> 143        ext                  TRUE              info  undocumented_file
#> 144        ext                  TRUE              info  undocumented_file
#> 145        ext                  TRUE              info  undocumented_file
#> 146        ext                  TRUE              info  undocumented_file
#> 147        ext                  TRUE              info  undocumented_file
#> 148        ext                  TRUE              info  undocumented_file
#> 149        ext                  TRUE              info  undocumented_file
#> 150        ext                  TRUE              info  undocumented_file
#> 151        ext                  TRUE              info  undocumented_file
#> 152        ext                  TRUE              info  undocumented_file
#> 153        ext                  TRUE              info  undocumented_file
#> 154        ext                  TRUE              info  undocumented_file
#> 155        ext                  TRUE              info  undocumented_file
#> 156        ext                  TRUE              info  undocumented_file
#> 157        ext                  TRUE              info  undocumented_file
#> 158        ext                  TRUE              info  undocumented_file
#> 159        ext                  TRUE              info  undocumented_file
#> 160        ext                  TRUE              info  undocumented_file
#> 161        ext                  TRUE              info  undocumented_file
#> 162        ext                  TRUE              info  undocumented_file
#> 163        ext                  TRUE              info  undocumented_file
#> 164        ext                  TRUE              info  undocumented_file
#> 165        ext                  TRUE              info  undocumented_file
#> 166        ext                  TRUE              info  undocumented_file
#> 167        ext                  TRUE              info  undocumented_file
#> 168        ext                  TRUE              info  undocumented_file
#> 169        ext                  TRUE              info  undocumented_file
#> 170        ext                  TRUE              info  undocumented_file
#> 171        ext                  TRUE              info  undocumented_file
#> 172        ext                  TRUE              info  undocumented_file
#> 173        ext                  TRUE              info  undocumented_file
#> 174        ext                  TRUE              info  undocumented_file
#> 175        ext                  TRUE              info  undocumented_file
#> 176        ext                  TRUE              info  undocumented_file
#> 177        ext                  TRUE              info  undocumented_file
#> 178        ext                  TRUE              info  undocumented_file
#> 179        ext                  TRUE              info  undocumented_file
#> 180        ext                  TRUE              info  undocumented_file
#> 181        ext                  TRUE              info  undocumented_file
#> 182        ext                  TRUE              info  undocumented_file
#> 183        ext                  TRUE              info  undocumented_file
#> 184        ext                  TRUE              info  undocumented_file
#> 185        ext                  TRUE              info  undocumented_file
#> 186        ext                  TRUE              info  undocumented_file
#> 187        ext                  TRUE              info  undocumented_file
#> 188        ext                  TRUE              info  undocumented_file
#> 189        ext                  TRUE              info  undocumented_file
#> 190        ext                  TRUE              info  undocumented_file
#> 191        ext                  TRUE              info  undocumented_file
#> 192        ext                  TRUE              info  undocumented_file
#> 193        ext                  TRUE              info  undocumented_file
#> 194        ext                  TRUE              info  undocumented_file
#> 195        ext                  TRUE              info  undocumented_file
#> 196        ext                  TRUE              info  undocumented_file
#> 197        ext                  TRUE              info  undocumented_file
#> 198        ext                  TRUE              info  undocumented_file
#> 199        ext                  TRUE              info  undocumented_file

gtfs$shapes <- NULL
validation_result <- validate_gtfs(gtfs)

# should raise a warning
gtfs$stop_times <- NULL
validation_result <- validate_gtfs(gtfs)

#> Warning: Invalid feed. Missing required file(s): stop_times.txt